refactor: make loadworkflows shared and establish a generic workflow-domain folder (#1528)
