+++ app/apis/catalog/brc-analytics-catalog/common/utils.ts
@@ -1,9 +1,9 @@
-import { formatTrsId } from "@/views/AnalyzeWorkflowsView/components/Main/utils";
 import {
   ORGANISM_PLOIDY,
   WORKFLOW_PLOIDY,
 } from "@repo/shared/apis/schema-types";
 import { sanitizeEntityId } from "@repo/shared/apis/utils";
+import { formatTrsId } from "@repo/shared/workflow/utils";
 import { WorkflowEntity } from "@site-config/brc-analytics/local/index/workflow/types";
 import { BRCDataCatalogGenome, BRCDataCatalogOrganism } from "./entities";
