+++ app/apis/catalog/ga2/utils.ts
@@ -1,5 +1,5 @@
-import { formatTrsId } from "@/views/AnalyzeWorkflowsView/components/Main/utils";
 import { sanitizeEntityId } from "@repo/shared/apis/utils";
+import { formatTrsId } from "@repo/shared/workflow/utils";
 import { WorkflowEntity } from "@site-config/ga2/local/index/workflow/types";
 import { GA2AssemblyEntity, GA2OrganismEntity } from "./entities";
