+++ app/views/AnalyzeWorkflowsView/components/Main/components/Accordion/components/Workflow/workflow.tsx
@@ -1,5 +1,4 @@
 import { TYPOGRAPHY_PROPS as COMPONENT_TYPOGRAPHY_PROPS } from "@/views/AnalyzeWorkflowsView/components/Main/components/Accordion/constants";
-import { formatTrsId } from "@/views/AnalyzeWorkflowsView/components/Main/utils";
 import { REL_ATTRIBUTE } from "@databiosphere/findable-ui/lib/components/Links/common/entities";
 import { Link as DXLink } from "@databiosphere/findable-ui/lib/components/Links/components/Link/link";
 import { BUTTON_PROPS } from "@databiosphere/findable-ui/lib/styles/common/mui/button";
@@ -7,6 +6,7 @@ import { TYPOGRAPHY_PROPS } from "@databiosphere/findable-ui/lib/styles/common/m
 import { replaceParameters } from "@databiosphere/findable-ui/lib/utils/replaceParameters";
 import { Button, Grid, Typography } from "@mui/material";
 import { ROUTES } from "@repo/shared/routes/constants";
+import { formatTrsId } from "@repo/shared/workflow/utils";
 import Link from "next/link";
 import { Fragment, JSX } from "react";
 import { GRID_PROPS } from "./constants";
