+++ packages/shared/workflow/utils.ts
@@ -1,0 +1,9 @@
+/**
+ * Formats a trsId for use in URLs by removing the hash character if it begins with one
+ * and replacing any special characters with hyphens.
+ * @param trsId - The trsId to format.
+ * @returns The formatted trsId.
+ */
+export function formatTrsId(trsId: string): string {
+  return trsId.replace(/^#/, "").replace(/[^a-zA-Z0-9]/g, "-");
+}
