+++ pages/data/[entityListType]/[entityId]/analyze/custom.tsx
@@ -2,9 +2,9 @@ import { getPageMeta } from "@/common/meta/utils";
 import { config } from "@/config/config";
 import { getEntities } from "@/utils/entityUtils";
 import { seedDatabase } from "@/utils/seedDatabase";
-import { CUSTOM_WORKFLOW } from "@/views/AnalyzeWorkflowsView/custom/constants";
 import { WorkflowInputsView } from "@/views/WorkflowInputsView/workflowInputsView";
 import { EntityDataGate } from "@repo/shared/components/EntityDataGate/entityDataGate";
+import { CUSTOM_WORKFLOW } from "@repo/shared/workflow/custom";
 import {
   GetStaticPaths,
   GetStaticPathsResult,
