+++ pages/data/workflows/[trsId]/index.tsx
@@ -1,11 +1,11 @@
 import { getPageMeta } from "@/common/meta/utils";
 import { config } from "@/config/config";
-import { formatTrsId } from "@/views/AnalyzeWorkflowsView/components/Main/utils";
-import { DIFFERENTIAL_EXPRESSION_ANALYSIS } from "@/views/AnalyzeWorkflowsView/differentialExpressionAnalysis/constants";
-import { LEXICMAP } from "@/views/AnalyzeWorkflowsView/lexicmap/constants";
-import { LOGAN_SEARCH } from "@/views/AnalyzeWorkflowsView/loganSearch/constants";
 import { WorkflowView } from "@/views/WorkflowView/workflowView";
 import { EntityDataGate } from "@repo/shared/components/EntityDataGate/entityDataGate";
+import { DIFFERENTIAL_EXPRESSION_ANALYSIS } from "@repo/shared/workflow/differentialExpressionAnalysis";
+import { LEXICMAP } from "@repo/shared/workflow/lexicmap";
+import { LOGAN_SEARCH } from "@repo/shared/workflow/loganSearch";
+import { formatTrsId } from "@repo/shared/workflow/utils";
 import { GetStaticPaths, GetStaticProps, GetStaticPropsContext } from "next";
 import { ParsedUrlQuery } from "querystring";
 import { JSX } from "react";
